fix: auto completer test dependency issue
